requirements.txt
amqp==5.0.2
APScheduler==3.6.3
asgiref==3.3.1
beautifulsoup4==4.9.3
billiard==3.6.3.0
celery==5.0.2
certifi==2020.11.8
cffi==1.14.3
chardet==3.0.4
click==7.1.2
click-didyoumean==0.0.3
click-repl==0.1.6
colorama==0.4.4
cryptography==3.2.1
decorator==4.4.2
Django==3.1.3
django-bootstrap4==2.3.1
django-celery-beat==2.1.0
django-cleanup==5.1.0
django-crispy-forms==1.10.0
django-timezone-field==4.0
django-widget-tweaks==1.4.8
idna==2.10
importlib-metadata==2.0.0
kombu==5.0.2
loguru==0.5.3
lxml==4.6.1
prompt-toolkit==3.0.8
pycparser==2.20
pyTelegramBotAPI==3.7.4
python-crontab==2.5.1
python-dateutil==2.8.1
python-decouple==3.3
python-telegram-bot==13.0
pytz==2020.4
redis==3.5.3
requests==2.25.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
vine==5.0.0
wcwidth==0.2.5
win32-setctime==1.0.3
zipp==3.4.0
